Add files via upload
After retrieving the number of scope in the source DHCP, each scope is exported. The name of the export xml file is constructed with the network ID of the scope. After the export, the number of file and the number of scope are compared. If the number is the same, the import is performed, otherwise a message appears and the script is stopped.

The script can be executed from the source or destination.